Refactor mr_tests.cpp to enable creating multithreaded tests that share the same infrastructure.
